feat: 상대방의 게시글 불러오는 로직 변경 페이징 처리 추가

+++ src/main/java/com/sparta/bluemoon/repository/PostRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
 
-    Long countByUserNot(User user);
+    int countByUserNot(User user);
 
     List<Post> findByUser(User user);
